refactor(cef): carve window transparency/opacity handlers into window/transparency.rs
Fourth sub-PR of the commands/window.rs modularization (Plan 1). Pure
move — no behavior change.

Extracted to window/transparency.rs: set_window_transparency,
set_window_opacity, get_window_opacity (pub, dispatched by ipc.rs,
re-exported `pub use transparency::*`) plus the private Win32 helpers
apply_window_opacity / remove_window_opacity (no external callers; used
only by the two setters, which both moved).

mod.rs dropped the `use lifecycle::find_all_own_windows` import — its
only consumer (set_window_transparency) moved out; transparency.rs now
imports it from the sibling lifecycle module under `#[cfg(windows)]`.
`capture_hwnd_for_label` re-export stays (external callers; never called
within mod.rs).

Verified: `cargo check -p agentmux-cef --release` clean.
