+++ .changesets/1780033446-refactor-cef-extract-window-transparency-opacity-handlers-in-dv2m.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): extract window transparency/opacity handlers into commands/window/transparency.rs
